modify: fix the owner logic

+++ Web/Backend/src/services/web/admin/owner-service.js
@@ -72,14 +72,21 @@ function createOwner(uid, userdata ,data, files) {
     return new Promise((resolve, reject) => {
         authHelper.hasOwnerPermission(userdata, [code.EDIT_PERMISSION]).then((response) => {
             ownerModel.createOwner_info(uid, data, files).then((response) => {
-                ownerModel.createBuildingRelationShip(data).then((response) => {
-                    ownerModel.createOwner(uid, data, response).then((response) => {
-                        let token = jwt.sign({ uid: uid, userdata: userdata }, key.JWT_SECRET_KEY, {
-                            expiresIn: timer.TOKEN_EXPIRATION
-                        })
-                        resolve({ code: code.OK, message: '', data: { 'token': token} })
+                if (data.owner_role === "subaccount") {
+                    let token = jwt.sign({ uid: uid, userdata: userdata }, key.JWT_SECRET_KEY, {
+                        expiresIn: timer.TOKEN_EXPIRATION
                     })
-                })
+                    resolve({ code: code.OK, message: '', data: { 'token': token} })
+                } else {
+                    ownerModel.createBuildingRelationShip(data).then((response) => {
+                        ownerModel.createOwner(uid, data, response).then((response) => {
+                            let token = jwt.sign({ uid: uid, userdata: userdata }, key.JWT_SECRET_KEY, {
+                                expiresIn: timer.TOKEN_EXPIRATION
+                            })
+                            resolve({ code: code.OK, message: '', data: { 'token': token} })
+                        })
+                    })
+                }
             }).catch((err) => {
                 if (err.message === message.INTERNAL_SERVER_ERROR)
                     reject({ code: code.INTERNAL_SERVER_ERROR, message: err.message, data: {} })
@@ -132,15 +139,21 @@ function updateOwner(uid, userdata ,data, files, id) {
     return new Promise((resolve, reject) => {
         authHelper.hasOwnerPermission(userdata, [code.EDIT_PERMISSION]).then((response) => {
             ownerModel.updateOwner_info(id, data, files).then((response) => {
-                ownerModel.delete_apartments(data, id).then((response) => {
-                    ownerModel.createOwner(uid, data, id).then((response) => {
-                        let token = jwt.sign({ uid: uid, userdata: userdata }, key.JWT_SECRET_KEY, {
-                            expiresIn: timer.TOKEN_EXPIRATION
-                        })
-                        resolve({ code: code.OK, message: '', data: { 'token': token} })
+                if (data.owner_role === "subaccount") {
+                    let token = jwt.sign({ uid: uid, userdata: userdata }, key.JWT_SECRET_KEY, {
+                        expiresIn: timer.TOKEN_EXPIRATION
                     })
-                })
-                
+                    resolve({ code: code.OK, message: '', data: { 'token': token} })
+                } else {
+                    ownerModel.delete_apartments(data, id).then((response) => {
+                        ownerModel.createOwner(uid, data, id).then((response) => {
+                            let token = jwt.sign({ uid: uid, userdata: userdata }, key.JWT_SECRET_KEY, {
+                                expiresIn: timer.TOKEN_EXPIRATION
+                            })
+                            resolve({ code: code.OK, message: '', data: { 'token': token} })
+                        })
+                    })
+                }
             }).catch((err) => {
                 if (err.message === message.INTERNAL_SERVER_ERROR)
                     reject({ code: code.INTERNAL_SERVER_ERROR, message: err.message, data: {} })
